+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Filter loans: Identifiers, created by and updated by columns [#4098]
 - Edit/New loan task: Add `none` status option
 - Edit/New Field occurence task: biological associations panel [#4103]
+- Edit/New Field occurence task: depictions panel [#4108]
 
 ### Fixed
 
@@ -24,6 +25,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#4098]: https://github.com/SpeciesFileGroup/taxonworks/issues/4098
 [#4100]: https://github.com/SpeciesFileGroup/taxonworks/issues/4100
 [#4103]: https://github.com/SpeciesFileGroup/taxonworks/issues/4103
+[#4108]: https://github.com/SpeciesFileGroup/taxonworks/issues/4108
 
 ## [0.45.0] - 2024-10-30